requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & Data Processing
anthropic==0.18.1
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2

# HTTP & Async
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
pydantic-core==2.14.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Monitoring & Logging
sentry-sdk==1.40.0
